+++ MAT-I/Assets/Scripts/ScriptableObjects/ItemDataScriptableObject.cs
@@ -31,9 +31,9 @@ public struct ItemData
     public Sprite icon;
     [TextArea]
     public string description;
-    public float buyingprice;
-    public float sellingprice;
-    public int weight;
+    public int buyingprice;
+    public int sellingprice;
+    public float weight;
     public ItemType itemType;
     public Rarity rarity;
     public int quantity;
